refactor: decouple api contracts from feign clients
