--- awesome-dto/src/main/java/tt/heixiong/awesome/api/ValidateApi.java
@@ -1,7 +1,0 @@
-package tt.heixiong.awesome.api;
-
-import org.springframework.cloud.openfeign.FeignClient;
-
-@FeignClient(value = "awesome")
-public interface ValidateApi {
-}
