+++ awesome-dto/src/main/java/tt/heixiong/awesome/contract/RequirementContract.java
@@ -1,6 +1,5 @@
-package tt.heixiong.awesome.api;
+package tt.heixiong.awesome.contract;
 
-import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import tt.heixiong.awesome.dto.RequirementDto;
@@ -9,9 +8,8 @@ import tt.heixiong.awesome.req.RequirementUpdateStatusReq;
 
 import java.util.List;
 
-@FeignClient(value = "awesome")
 @RequestMapping("/requirements")
-public interface RequirementApi {
+public interface RequirementContract {
 
     @PostMapping
     RequirementDto createRequirement(@Validated @RequestBody RequirementCreateReq req);
