+++ awesome-web/src/main/java/tt/heixiong/awesome/client/RequirementClient.java
@@ -1,0 +1,8 @@
+package tt.heixiong.awesome.client;
+
+import org.springframework.cloud.openfeign.FeignClient;
+import tt.heixiong.awesome.contract.RequirementContract;
+
+@FeignClient(value = "awesome")
+public interface RequirementClient extends RequirementContract {
+}
